# Various DNN improvements (rebase PR61, PR100, PR103, and PR114)