--- a/RecoTracker/LSTCore/interface/QuintupletsSoA.h
+++ b/RecoTracker/LSTCore/interface/QuintupletsSoA.h
@@ -28,7 +28,9 @@ namespace lst {
                       SOA_COLUMN(float, regressionF),
                       SOA_COLUMN(float, rzChiSquared),  // r-z only chi2
                       SOA_COLUMN(float, chiSquared),
-                      SOA_COLUMN(float, nonAnchorChiSquared));
+                      SOA_COLUMN(float, nonAnchorChiSquared),
+                      SOA_COLUMN(float, dBeta1),
+                      SOA_COLUMN(float, dBeta2));
 
   using QuintupletsSoA = QuintupletsSoALayout<>;
   using Quintuplets = QuintupletsSoA::View;