--- a/RecoTracker/LSTCore/src/alpaka/NeuralNetwork.h
+++ b/RecoTracker/LSTCore/src/alpaka/NeuralNetwork.h
@@ -38,6 +38,18 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE::lst::t5dnn {
     }
   }
 
+  ALPAKA_FN_ACC ALPAKA_FN_INLINE float delta_phi(const float phi1, const float phi2) {
+    float delta = phi1 - phi2;
+    // Adjust delta to be within the range [-M_PI, M_PI]
+    if (delta > kPi) {
+      delta -= 2 * kPi;
+    } else if (delta < -kPi) {
+      delta += 2 * kPi;
+    }
+
+    return delta;
+  }
+
   template <typename TAcc>
   ALPAKA_FN_ACC ALPAKA_FN_INLINE bool runInference(TAcc const& acc,
                                                    MiniDoubletsConst mds,
@@ -50,7 +62,7 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE::lst::t5dnn {
                                                    const float outerRadius,
                                                    const float bridgeRadius) {
     // Constants
-    constexpr unsigned int kinputFeatures = 18;
+    constexpr unsigned int kinputFeatures = 23;
     constexpr unsigned int khiddenFeatures = 32;
 
     float eta1 = alpaka::math::abs(acc, mds.anchorEta()[mdIndex1]);  // inner T3 anchor hit 1 eta (t3_0_eta)
@@ -58,6 +70,12 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE::lst::t5dnn {
     float eta3 = alpaka::math::abs(acc, mds.anchorEta()[mdIndex3]);  // inner T3 anchor hit 3 eta (t3_4_eta)
     float eta4 = alpaka::math::abs(acc, mds.anchorEta()[mdIndex4]);  // outer T3 anchor hit 4 eta (t3_2_eta)
     float eta5 = alpaka::math::abs(acc, mds.anchorEta()[mdIndex5]);  // outer T3 anchor hit 5 eta (t3_4_eta)
+
+    float phi1 = mds.anchorPhi()[mdIndex1];  // inner T3 anchor hit 1 phi (t3_0_phi)
+    float phi2 = mds.anchorPhi()[mdIndex2];  // inner T3 anchor hit 2 phi (t3_2_phi)
+    float phi3 = mds.anchorPhi()[mdIndex3];  // inner T3 anchor hit 3 phi (t3_4_phi)
+    float phi4 = mds.anchorPhi()[mdIndex4];  // outer T3 anchor hit 4 phi (t3_2_phi)
+    float phi5 = mds.anchorPhi()[mdIndex5];  // outer T3 anchor hit 5 phi (t3_4_phi)
 
     float z1 = alpaka::math::abs(acc, mds.anchorZ()[mdIndex1]);  // inner T3 anchor hit 1 z (t3_0_z)
     float z2 = alpaka::math::abs(acc, mds.anchorZ()[mdIndex2]);  // inner T3 anchor hit 2 z (t3_2_z)
@@ -73,25 +91,30 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE::lst::t5dnn {
 
     // Build the input feature vector using pairwise differences after the first hit
     float x[kinputFeatures] = {
-        eta1 / kEta_norm,  // inner T3: First hit eta normalized
-        z1 / kZ_max,       // inner T3: First hit z normalized
-        r1 / kR_max,       // inner T3: First hit r normalized
+        eta1 / kEta_norm,                          // inner T3: First hit eta normalized
+        alpaka::math::abs(acc, phi1) / kPhi_norm,  // inner T3: First hit phi normalized
+        z1 / kZ_max,                               // inner T3: First hit z normalized
+        r1 / kR_max,                               // inner T3: First hit r normalized
 
-        eta2 - eta1,         // inner T3: Difference in eta between hit 2 and 1
-        (z2 - z1) / kZ_max,  // inner T3: Difference in z between hit 2 and 1 normalized
-        (r2 - r1) / kR_max,  // inner T3: Difference in r between hit 2 and 1 normalized
+        eta2 - eta1,                        // inner T3: Difference in eta between hit 2 and 1
+        delta_phi(phi2, phi1) / kPhi_norm,  // inner T3: Difference in phi between hit 2 and 1
+        (z2 - z1) / kZ_max,                 // inner T3: Difference in z between hit 2 and 1 normalized
+        (r2 - r1) / kR_max,                 // inner T3: Difference in r between hit 2 and 1 normalized
 
-        eta3 - eta2,         // inner T3: Difference in eta between hit 3 and 2
-        (z3 - z2) / kZ_max,  // inner T3: Difference in z between hit 3 and 2 normalized
-        (r3 - r2) / kR_max,  // inner T3: Difference in r between hit 3 and 2 normalized
+        eta3 - eta2,                        // inner T3: Difference in eta between hit 3 and 2
+        delta_phi(phi3, phi2) / kPhi_norm,  // inner T3: Difference in phi between hit 3 and 2
+        (z3 - z2) / kZ_max,                 // inner T3: Difference in z between hit 3 and 2 normalized
+        (r3 - r2) / kR_max,                 // inner T3: Difference in r between hit 3 and 2 normalized
 
-        eta4 - eta3,         // outer T3: Difference in eta between hit 4 and 3
-        (z4 - z3) / kZ_max,  // outer T3: Difference in z between hit 4 and 3 normalized
-        (r4 - r3) / kR_max,  // outer T3: Difference in r between hit 4 and 3 normalized
+        eta4 - eta3,                        // outer T3: Difference in eta between hit 4 and 3
+        delta_phi(phi4, phi3) / kPhi_norm,  // inner T3: Difference in phi between hit 4 and 3
+        (z4 - z3) / kZ_max,                 // outer T3: Difference in z between hit 4 and 3 normalized
+        (r4 - r3) / kR_max,                 // outer T3: Difference in r between hit 4 and 3 normalized
 
-        eta5 - eta4,         // outer T3: Difference in eta between hit 5 and 4
-        (z5 - z4) / kZ_max,  // outer T3: Difference in z between hit 5 and 4 normalized
-        (r5 - r4) / kR_max,  // outer T3: Difference in r between hit 5 and 4 normalized
+        eta5 - eta4,                        // outer T3: Difference in eta between hit 5 and 4
+        delta_phi(phi5, phi4) / kPhi_norm,  // inner T3: Difference in phi between hit 5 and 4
+        (z5 - z4) / kZ_max,                 // outer T3: Difference in z between hit 5 and 4 normalized
+        (r5 - r4) / kR_max,                 // outer T3: Difference in r between hit 5 and 4 normalized
 
         alpaka::math::log10(acc, innerRadius),   // T5 inner radius (t5_innerRadius)
         alpaka::math::log10(acc, bridgeRadius),  // T5 bridge radius (t5_bridgeRadius)