This is a rebase of PRs #61, #100, #103, and #114. They were all DNN-related, so I grouped them together.

One thing I should mention is that I replaced `#pragma unroll N` with `CMS_UNROLL_LOOP`. There is a `CMS_UNROLL_LOOP_COUNT(N)` macro, but it only works with explicit values of `N`. However, it seems like `CMS_UNROLL_LOOP` works equally well as `#pragma unroll N` and we avoid the warnings.