@slava77 2024-11-27
will this be squash-merged?
It's not obvious we need the history of the notebooks and the network weights