@ariostas 2024-11-28
> will this be squash-merged?
> It's not obvious we need the history of the notebooks and the network weights

Yeah, this one makes sense to squash-merge